Move `fallbackLocale` function to the i18n module
